+++ alpha/lib/model/FileSync.php
@@ -282,7 +282,7 @@ class FileSync extends BaseFileSync implements IBaseObject
 	{
 		return (isset($this->statusMap[$this->getStatus()])) ? $this->statusMap[$this->getStatus()] : "Unknown";
 	}
-	
+
 	public function getExternalUrl($entryId, $format = PlaybackProtocol::HTTP)
 	{
 		$storage = StorageProfilePeer::retrieveByPK($this->getDc());
@@ -293,23 +293,46 @@ class FileSync extends BaseFileSync implements IBaseObject
 		{
 			return null;
 		}
+		$kalturaPeriodicStorage = false;
+		if(in_array($this->getDc(), kStorageExporter::getPeriodicStorageIdsByPartner($this->getPartnerId())))
+		{
+			$kalturaPeriodicStorage = true;
+		}
 
 		$urlManager = DeliveryProfilePeer::getRemoteDeliveryByStorageId(DeliveryProfileDynamicAttributes::init($this->getDc(), $entryId, PlaybackProtocol::HTTP, infraRequestUtils::getProtocol()));
-		if(is_null($urlManager) && infraRequestUtils::getProtocol() != 'http')
+		if(is_null($urlManager) && infraRequestUtils::getProtocol() != 'http' && !$kalturaPeriodicStorage)
 			$urlManager = DeliveryProfilePeer::getRemoteDeliveryByStorageId(DeliveryProfileDynamicAttributes::init($this->getDc(), $entryId));
 		if(is_null($urlManager))
 			return null;
 		
 		$url = $urlManager->getFileSyncUrl($this);
 		$baseUrl = $urlManager->getUrl();
-		
+
+		if($kalturaPeriodicStorage)
+		{
+			$url = '/direct' . $url;
+			$authParams = $this->addKalturaAuthParams($url);
+			$url .= $authParams;
+		}
+
 		$url = ltrim($url, "/");
-		if (strpos($url, "://") === false){
+		if (strpos($url, "://") === false)
+		{
 			$url = rtrim($baseUrl, "/") . "/".$url ;
 		}
+
 		return $url;
 	}
-	
+
+	protected function addKalturaAuthParams($url)
+	{
+		$version = kNetworkUtils::DEFAULT_AUTH_HEADER_VERSION;
+		$timestamp = time();
+		$signature = kNetworkUtils::calculateSignature($version, $timestamp, $url);
+
+		return '?kaltura_auth=' . $version . ',' . $timestamp . ',' . $signature;
+	}
+
 	/* (non-PHPdoc)
 	 * @see BaseFileSync::preUpdate()
 	 */