+++ deployment/updates/scripts/2020_06_14_deploy_caption_asset_boolean_notification.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @package deployment
+ */
+require_once (__DIR__ . '/../../bootstrap.php');
+$script = realpath(dirname(__FILE__) . "/../../../tests/standAloneClient/exec.php");
+$newTemplateUpdate = realpath(dirname(__FILE__) . "/../../updates/scripts/xml/2020_06_14_captionAssetBooleanEventNotification.xml");
+print_r($newTemplateUpdate);
+if(!file_exists($newTemplateUpdate) || !file_exists($script))
+{
+	KalturaLog::err("Missing update script file");
+	return;
+}
+passthru("php $script $newTemplateUpdate");