+++ deployment/updates/scripts/add_permissions/2020_05_28_add_permission_analytics_base_to_tag_search.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @package deployment
+ *
+ * Adding permission ANALYTICS_BASE to search action in tagsearch_tag
+ */
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$config = realpath(dirname(__FILE__)) . '/../../../permissions/service.tagsearch.tag.ini';
+passthru("php $script $config");