+++ plugins/search/providers/elastic_search/lib/model/enum/ESearchEntryFilterFields.php
@@ -50,6 +50,7 @@ class ESearchEntryFilterFields extends KalturaEnum
 	const PARTNER_SORT_VALUE = 'partner_sort_value';
 	const SEARCH_TEXT = 'search_text';
 	const FREE_TEXT	= '_free_text';
+	const EXCLUDED_FREE_TEXT_GROUPS	= '_excluded_free_text_groups';
 	const TOTAL_RANK = 'total_rank';
 	const RANK = 'rank';
 	const VOTES = 'votes';